% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_citihub.R
\name{theme_citihub}
\alias{theme_citihub}
\title{\code{theme_citihub} A citihubernment theme for \code{ggplot2}}
\usage{
theme_citihub(axes = "x")
}
\arguments{
\item{axes}{Character string. Specifies the presence or absence of axes
Must be one of \code{n} (no axes), \code{x} (only x axis), \code{y}
(only y axis), or \code{xy} (both axes shown).}

\item{base_size}{Integer. Sets the base size of text for the plot.
Defaults to \code{12}.}

\item{base_colour}{Character string. Sets the default colour of axes
 and axis labels. Must be a named R colour or hexadecimal colour code
(e.g. "#FF0000"). Defaults to \code{gray40}.}
}
\value{
Will not return anything of itself, but when used in conjuntion
with \code{\link{ggplot}} and (e.g.) \code{\link{geom_point}} from the
package \code{ggplot2}, will apply styling to a plot.
}
\description{
\code{theme_citihub} Provides a theme for ggplot2 to produce
Citihub style visualisations in \code{ggplot2}.
}
\details{
Builds on the 'grammar of graphics' framework implement in
ggplot2. Applying \code{theme_citihub()} will adjust graphical parameters
to give a plot a feel more in line with citihub.com.
}
\examples{


library(ggplot2)
library(citihubstyle)
library(dplyr)

p <- mtcars \%>\%
   ggplot +
   aes(
   x = mpg,
   y = wt
   ) +
   geom_point()

# Plot without any theme applied

p

# Now apply citihub theme

p +
 theme_citihub()

}
